requirements.txt
aiohttp==3.8.4
async-timeout==4.0.2
attrs==22.2.0
chardet==5.1.0
discord.py==2.1.1
python-dotenv==1.0.0
idna==3.4
multidict==6.0.4
typing-extensions==4.5.0
yarl==1.8.2